// app/models/User.ts
export class User{
    idRukovodioca: number = 0;
    username: string = "";
    lozinka: string = "";
    ime: string = "";
    prezime: string = "";
    datumRodjenja!: Date;
    mailRukovodioca: string = "";
    telefon: string = "";
    pregledPodnetih: string = "";
    idInstitucije!: number;
}